requirements.txt
beautifulsoup4==4.3.1
djangorestframework>=2.3.14
